GRAPHICS-8: Improve error handling
On the client-side, properly show the message and status code (and no more "oopsie woopsie"...) when there's a fetch error or the response doesn't have the status code we expected.

On the server side, wrap all our handlers in express-async-handler to ensure any asynchronous errors get caught (desperately waiting for Express 5.0...), as well as slight improvements to 404s and logging.